Finished prototype

+++ app/src/main/java/csusb/cse541/william/cse541robotics/Constants.java
@@ -2,16 +2,18 @@ package csusb.cse541.william.cse541robotics;
 
 /**
  * Created by William on 5/20/2015.
+ *
  */
+
 public class Constants {
     /* These are the movement signals that are sent to the robot */
-    public final static String FORWARD_SIG= "f";
-    public final static String BACKWARD_SIG= "b";
+    public final static String FORWARD_SIG = "f";
+    public final static String BACKWARD_SIG = "b";
     public final static String LEFT_SIG = "l";
     public final static String RIGHT_SIG = "r";
     public final static String STOP_SIG = "s";
 
     /* The settings for the wifi connection */
     public final static String REMOTE_IP_ADDRESS = "";
-    public final static short REMOTE_PORT = 23;
+    public final static int REMOTE_PORT = 23;
 }
